+++ app/src/main/java/com/example/android/ballBounce/paintableShapes/PaintableCircle.kt
@@ -4,22 +4,22 @@ import android.graphics.Canvas
 import com.example.android.ballBounce.utility.Vector
 import com.example.android.ballBounce.view.CirclePaintFactory
 
-class PaintableCircle(val center: Vector, val radius: Float,val colorIndex: Int = 0) : PaintableShape() {
+class PaintableCircle(val center: Vector, val radius: Double, val colorIndex: Int = 0) : PaintableShape() {
 
     lateinit var circlePaintFactory: CirclePaintFactory
 
     override fun paintShape(paintCanvas: Canvas) {
         paintCanvas.drawCircle(
-            center.x,
-            center.y,
-            radius,
+            center.x.toFloat(),
+            center.y.toFloat(),
+            radius.toFloat(),
             circlePaintFactory.getFillPaint(colorIndex)
         )
 
         paintCanvas.drawCircle(
-            center.x,
-            center.y,
-            radius-circlePaintFactory.getOutlinePaint().strokeWidth/2,
+            center.x.toFloat(),
+            center.y.toFloat(),
+            (radius-circlePaintFactory.getOutlinePaint().strokeWidth/2).toFloat(),
             circlePaintFactory.getOutlinePaint()
         )
     }
