+++ app/src/main/java/com/example/android/ballBounce/paintableShapes/PaintableLine.kt
@@ -4,11 +4,11 @@ import android.graphics.Canvas
 import com.example.android.ballBounce.utility.Vector
 import com.example.android.ballBounce.view.LinePaintFactory
 
-class PaintableLine(val start: Vector, val end: Vector, val width: Float) : PaintableShape() {
+class PaintableLine(val start: Vector, val end: Vector, val width: Double) : PaintableShape() {
     lateinit var linePaintFactory: LinePaintFactory
 
     override fun paintShape(paintCanvas: Canvas) {
-        paintCanvas.drawLine(start.x,start.y,end.x,end.y,linePaintFactory.getPaint(width))
+        paintCanvas.drawLine(start.x.toFloat(),start.y.toFloat(),end.x.toFloat(),end.y.toFloat(),linePaintFactory.getPaint(width))
     }
 
 }
