+++ app/src/main/java/com/example/android/ballBounce/paintableShapes/PaintableText.kt
@@ -8,6 +8,6 @@ class PaintableText(val text:String,val position: Vector): PaintableShape() {
     lateinit var textPaintFactory: TextPaintFactory
 
     override fun paintShape(paintCanvas: Canvas) {
-        paintCanvas.drawText(text,position.x,position.y,textPaintFactory.getPaint())
+        paintCanvas.drawText(text,position.x.toFloat(),position.y.toFloat(),textPaintFactory.getPaint())
     }
 }
