+++ app/src/main/java/com/example/android/ballBounce/view/LinePaintFactory.kt
@@ -3,9 +3,9 @@ package com.example.android.ballBounce.view
 import android.graphics.Paint
 
 class LinePaintFactory(private val strokeColor: Int) {
-    private var paintCache: HashMap<Float,Paint> = HashMap<Float,Paint>()
+    private var paintCache: HashMap<Double,Paint> = HashMap<Double,Paint>()
 
-    fun getPaint(strokeWidth: Float): Paint {
+    fun getPaint(strokeWidth: Double): Paint {
         if (paintCache.containsKey(strokeWidth)) {
             return paintCache[strokeWidth]!!
         }
@@ -15,7 +15,7 @@ class LinePaintFactory(private val strokeColor: Int) {
             isDither = true
             style = Paint.Style.STROKE
         }
-        paintCache[strokeWidth]!!.strokeWidth = strokeWidth
+        paintCache[strokeWidth]!!.strokeWidth = strokeWidth.toFloat()
         return paintCache[strokeWidth]!!
     }
 }
