+++ app/src/main/java/com/example/android/ballBounce/viewModel/MainActivityViewModel.kt
@@ -22,7 +22,7 @@ class MainActivityViewModel() : ViewModel() {
         entitySim!!.initialize(screenDims)
     }
 
-    fun stepModel(dt: Float): Unit {
+    fun stepModel(dt: Double): Unit {
         entitySim?.updateState(dt)
     }
